Pull

+++ web/api/src/main/java/com/hergomsoft/easyoapi/controllers/CarrerasController.java
@@ -2,16 +2,30 @@ package com.hergomsoft.easyoapi.controllers;
 
 
 import com.hergomsoft.easyoapi.models.Carrera;
+import com.hergomsoft.easyoapi.models.Recorrido;
 import com.hergomsoft.easyoapi.models.Usuario;
+import com.hergomsoft.easyoapi.models.requests.UbicacionRequest;
+import com.hergomsoft.easyoapi.models.responses.CarreraSimplificada;
 import com.hergomsoft.easyoapi.services.CarreraService;
 import com.hergomsoft.easyoapi.services.UsuarioService;
-import java.net.URI;
-import java.util.HashMap;
+import java.io.ByteArrayInputStream;
+import java.io.IOException;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
+import java.util.TreeMap;
+import java.util.logging.Level;
+import java.util.logging.Logger;
+import java.util.zip.ZipEntry;
+import java.util.zip.ZipOutputStream;
+import javax.servlet.http.HttpServletResponse;
+import org.apache.tomcat.util.http.fileupload.IOUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
+import org.springframework.http.MediaType;
+import org.springframework.lang.Nullable;
+import org.springframework.util.StreamUtils;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -19,10 +33,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 @RestController
 @RequestMapping("/api/carreras")
@@ -34,18 +48,93 @@ public class CarrerasController {
     @Autowired
     private UsuarioService usuariosService;
     
-    @GetMapping("")
-    public List<Carrera> getCarreras() {
-        return carrerasService.findAll();
+    @GetMapping("/participadas")
+    public List<Carrera> getCarrerasParticipadasUsuario() {
+        Usuario usuario = usuariosService.getUsuarioPeticion();
+        List<Carrera> participadas = carrerasService.getCarrerasParticipadasUsuario(usuario);
+        return participadas;
+    }
+    
+    @GetMapping("/organizadas")
+    public List<Carrera> getCarrerasOrganizadasUsuario() {
+        Usuario usuario = usuariosService.getUsuarioPeticion();
+        List<Carrera> organizadas = carrerasService.getCarrerasOrganizadasUsuario(usuario);
+        return organizadas;
+    }
+    
+    @GetMapping("/buscar")
+    public List<CarreraSimplificada> buscaCarreras(
+            @Nullable @RequestParam("nombre") String nombre, 
+            @Nullable @RequestParam("tipo") String tipo, 
+            @Nullable @RequestParam("modalidad") String modalidad,
+            Pageable pageable) {
+        if(nombre == null) nombre = "";
+        if(tipo == null) tipo = "";
+        if(modalidad == null) modalidad = "";
+        long idUsuario = usuariosService.getUsuarioPeticion().getId();
+        return carrerasService.buscaCarreras(idUsuario, nombre, tipo, modalidad, pageable);
+    }
+    
+    @GetMapping("/{idCarrera}/mapas")
+    public void getMapasCarrera(@PathVariable long idCarrera, HttpServletResponse response) {
+        // TODO Restricción solo usuario organizador
+        Carrera carrera = getCarrera(idCarrera);
+        response.setContentType("application/zip");
+        response.setHeader("Content-Disposition", "attachment; filename=mapas.zip");
+
+        try (ZipOutputStream zippedOut = new ZipOutputStream(response.getOutputStream())) {
+            for(Recorrido r : carrera.getRecorridos()) {
+                ZipEntry e = new ZipEntry(r.getNombre() + ".jpg");
+                byte[] mapa = r.getMapa();
+                e.setSize(mapa.length);
+                e.setTime(System.currentTimeMillis());
+                zippedOut.putNextEntry(e);
+                // And the content of the resource:
+                StreamUtils.copy(mapa, zippedOut);
+                zippedOut.closeEntry();
+            }
+            
+            zippedOut.finish();
+        } catch (Exception e) {
+            // Do something with Exception
+        }   
+    }
+    
+    @GetMapping("/mapa/{idRecorrido}")
+    public void getMapaRecorrido(@PathVariable long idRecorrido, HttpServletResponse response) {
+        // TODO Restricciones de descarga (solo participantes, organizador...)
+        Recorrido rec = carrerasService.getRecorrido(idRecorrido);
+        if(rec != null) {
+            if(rec.getMapa() != null) {
+                try {
+                    response.setContentType(MediaType.IMAGE_JPEG_VALUE);
+                    IOUtils.copy(new ByteArrayInputStream(rec.getMapa()), response.getOutputStream());
+                } catch (IOException ex) {
+                    Logger.getLogger(CarrerasController.class.getName()).log(Level.SEVERE, null, ex);
+                    throw new ResponseStatusException(
+                        HttpStatus.INTERNAL_SERVER_ERROR, "Error al obtener el mapa");
+                }
+            } else {
+                // No tiene mapa -> 404
+                throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, "El recorrido no tiene mapa");
+            }
+        } else {
+            // No existe -> 404
+            throw new ResponseStatusException(
+                HttpStatus.NOT_FOUND, "No existe ningún recorrido con ese ID");
+        }
+        
     }
     
     @GetMapping("/{id}")
     public Carrera getCarrera(@PathVariable long id) {
+        // TODO Si es privada solo pueden verlas los participantes y el organizador
         Carrera res = carrerasService.getCarrera(id);
         if(res == null) {
             // No existe -> 404
             throw new ResponseStatusException(
-                HttpStatus.NOT_FOUND, "No existe la carrera con ese ID");
+                HttpStatus.NOT_FOUND, "No existe ninguna carrera con ese ID");
         } else {
             return res;
         }
@@ -55,46 +144,55 @@ public class CarrerasController {
     @ResponseStatus(HttpStatus.CREATED)
     public Carrera nuevaCarrera(@RequestBody Carrera carrera) {
         // El creador de la carrera es el usuario que realiza la petición
-        // TODO
-        Usuario org = usuariosService.getUsuario(2L); // Usuario de prueba
+        //Usuario org = usuariosService.getUsuarioPeticion();
+        Usuario org = usuariosService.getUsuario(2L); // Usuario de prueba de creación
         carrera.setOrganizador(org);
-        // TODO Comprobar datos válidos
         
         // Crea la carrera
         return carrerasService.saveCarrera(carrera);
     }
     
     @PutMapping("/{id}")
-    public ResponseEntity<Object> editarCarrera(@RequestBody Carrera carrera, @PathVariable long id) {
-        Carrera res = carrerasService.getCarrera(id);
-        if(res == null) {
-            return ResponseEntity.notFound().build();
+    public void editarCarrera(@RequestBody Carrera carrera, @PathVariable long id) {
+        Usuario usuario = usuariosService.getUsuarioPeticion();
+        Carrera c = getCarrera(id);
+        //if(Objects.equals(usuario.getId(), c.getOrganizador().getId())) {
+        if(true) { // DEBUG
+            carrerasService.editCarrera(c, carrera);
+        } else {
+            // No es el organizador de la carrera
+            throw new ResponseStatusException(
+                HttpStatus.FORBIDDEN, "Solo el organizador puede editar la carrera");
         }
-
-        // TODO Comprobar datos válidos
-        carrera.setId(id);
-        carrerasService.editCarrera(carrera);
-        return ResponseEntity.noContent().build();
     }
     
     @DeleteMapping("/{id}")
     public void borrarCarrera(@PathVariable long id) {
-        carrerasService.deleteCarrera(id);
+        Usuario usuario = usuariosService.getUsuarioPeticion();
+        Carrera c = getCarrera(id);
+        if(Objects.equals(usuario.getId(), c.getOrganizador().getId())) {
+            carrerasService.deleteCarrera(id);
+        } else {
+            // No es el organizador de la carrera
+            throw new ResponseStatusException(
+                HttpStatus.FORBIDDEN, "Solo el organizador puede borrar la carrera.");
+        }
     }
     
-    @GetMapping("/{id}/secretos")
-    public Map<String, String> getSecretosControlesCarrera(@PathVariable long id) {
-        // Solo es accesible por el organizador de la carrera
-        // TODO 
+    @GetMapping("/{id}/qr")
+    public Map<String, String> getControlesQRCarrera(@PathVariable long id) {
+        Map<String, String> res = new TreeMap<>();
         
-        Map<String, String> res = new HashMap<>();
-        Carrera carrera = carrerasService.getCarrera(id);
-        if(carrera != null) {
-            res = carrerasService.getSecretosCarrera(carrera);
+        Usuario usuario = usuariosService.getUsuarioPeticion();
+        Carrera carrera = getCarrera(id);
+        // Solo es accesible por el organizador de la carrera
+        //if(Objects.equals(usuario.getId(), carrera.getOrganizador().getId())) {
+        if(true) { // DEBUG
+            res = carrerasService.getControlesConSecretosCarrera(carrera);
         } else {
-            // No existe -> 404
+            // No es el organizador de la carrera
             throw new ResponseStatusException(
-                HttpStatus.NOT_FOUND, "No existe la carrera con ese ID");
+                HttpStatus.FORBIDDEN, "Solo puede acceder el organizador de la carrera");
         }
         
         return res;
